scraper/requirements.txt
google-cloud-firestore==1.2.0
pyppeteer==0.0.25
python-dotenv==0.10.2
requests==2.21.0
SQLAlchemy==1.3.3
stringcase==1.2.0
urllib3==1.24.2
websockets==6.0
